refactor(validator): 메서드 분할
